💄 cosmetic(VSecM): code refactoring
