+++ app/safe/internal/server/route/audit.go
@@ -1,0 +1,31 @@
+/*
+|    Protect your secrets, protect your sensitive data.
+:    Explore VMware Secrets Manager docs at https://vsecm.com/
+</
+<>/  keep your secrets… secret
+>/
+<>/' Copyright 2023–present VMware Secrets Manager contributors.
+>/'  SPDX-License-Identifier: BSD-2-Clause
+*/
+
+package route
+
+import (
+	"net/http"
+	
+	"github.com/vmware-tanzu/secrets-manager/core/audit"
+	event "github.com/vmware-tanzu/secrets-manager/core/audit/state"
+	reqres "github.com/vmware-tanzu/secrets-manager/core/entity/reqres/safe/v1"
+)
+
+func createDefaultJournalEntry(cid, spiffeid string,
+	r *http.Request) audit.JournalEntry {
+	return audit.JournalEntry{
+		CorrelationId: cid,
+		Entity:        reqres.SecretFetchRequest{},
+		Method:        r.Method,
+		Url:           r.RequestURI,
+		SpiffeId:      spiffeid,
+		Event:         event.Enter,
+	}
+}
