+++ app/safe/internal/server/route/http_test.go
@@ -8,20 +8,4 @@
 >/'  SPDX-License-Identifier: BSD-2-Clause
 */
 
-package state
-
-import "strings"
-
-func ageKeyTriplet() (string, string, string) {
-	if rootKey == "" {
-		return "", "", ""
-	}
-
-	parts := strings.Split(rootKey, "\n")
-
-	if len(parts) != 3 {
-		return "", "", ""
-	}
-
-	return parts[0], parts[1], parts[2]
-}
+package route
